feat: added map generation

+++ nodo.py
@@ -1,12 +1,13 @@
 from typing import Tuple
 
+
 class Node:
     def __init__(
         self,
         position: Tuple[int, int],
         gas_pumps: int = 0,
         energy_chargers: int = 0,
-        energy_recharge_rate_km_h: float = 0.0,
+        energy_recharge_rate_km_h: int = 0.0,
     ) -> None:
         self.position = position
         # Atributos de Combustão
